Update format strings to take signedness into account.
The latest version of Clang 19 detects sign mismatches in format
strings when -Wformat is enabled. This CL corrects a majority of
format strings throughout our code, but may miss some as I don't
have every backend enabled.

Disappointingly, there is no format string representing "a signed
integer in hexadecimal," so these require casts to uint32_t.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/842336
Commit-Queue: John Stiles <johnstiles@google.com>

+++ src/base/SkSharedMutex.cpp
@@ -11,6 +11,7 @@
 #include "include/private/base/SkSemaphore.h"
 
 #include <cinttypes>
+#include <cstdint>
 
 #if !defined(__has_feature)
     #define __has_feature(x) 0
@@ -134,10 +135,11 @@
             SkAutoMutexExclusive l(fMu);
 
             SkASSERTF(!fCurrentShared->find(threadID),
-                      "Thread %" PRIx64 " already has an shared lock\n", threadID);
+                      "Thread %" PRIx64 " already has an shared lock\n", (uint64_t)threadID);
 
             if (!fWaitingExclusive->tryAdd(threadID)) {
-                SkDEBUGFAILF("Thread %" PRIx64 " already has an exclusive lock\n", threadID);
+                SkDEBUGFAILF("Thread %" PRIx64 " already has an exclusive lock\n",
+                             (uint64_t)threadID);
             }
 
             currentSharedCount = fCurrentShared->count();
@@ -164,7 +166,8 @@
             SkAutoMutexExclusive l(fMu);
             SkASSERT(0 == fCurrentShared->count());
             if (!fWaitingExclusive->tryRemove(threadID)) {
-                SkDEBUGFAILF("Thread %" PRIx64 " did not have the lock held.\n", threadID);
+                SkDEBUGFAILF("Thread %" PRIx64 " did not have the lock held.\n",
+                             (uint64_t)threadID);
             }
             exclusiveWaitingCount = fWaitingExclusive->count();
             sharedWaitingCount = fWaitingShared->count();
@@ -198,11 +201,12 @@
             exclusiveWaitingCount = fWaitingExclusive->count();
             if (exclusiveWaitingCount > 0) {
                 if (!fWaitingShared->tryAdd(threadID)) {
-                    SkDEBUGFAILF("Thread %" PRIx64 " was already waiting!\n", threadID);
+                    SkDEBUGFAILF("Thread %" PRIx64 " was already waiting!\n", (uint64_t)threadID);
                 }
             } else {
                 if (!fCurrentShared->tryAdd(threadID)) {
-                    SkDEBUGFAILF("Thread %" PRIx64 " already holds a shared lock!\n", threadID);
+                    SkDEBUGFAILF("Thread %" PRIx64 " already holds a shared lock!\n",
+                                 (uint64_t)threadID);
                 }
             }
             sharedQueueSelect = fSharedQueueSelect;
@@ -224,7 +228,8 @@
         {
             SkAutoMutexExclusive l(fMu);
             if (!fCurrentShared->tryRemove(threadID)) {
-                SkDEBUGFAILF("Thread %" PRIx64 " does not hold a shared lock.\n", threadID);
+                SkDEBUGFAILF("Thread %" PRIx64 " does not hold a shared lock.\n",
+                             (uint64_t)threadID);
             }
             currentSharedCount = fCurrentShared->count();
             waitingExclusiveCount = fWaitingExclusive->count();
